Fix #440: get the unit test suite to zero errors/failures on SQLite
After #440's first-pass reparenting, the Code Coverage unit step on
SQLite exposed 26 errors and 20 failures. Most were not regressions but
pre-existing bugs and test mistakes that the "no such table" errors had
been masking: once the migrations actually ran, the real shape of each
test showed up.

Library fixes (correct code, not test accommodations):
- MessageService: the fallback of `workflow_run_id` to the target
  instance id when an instance had no `current_run_id` wrote an
  instance id into a run FK column. SQLite with FK on rejected it as
  a constraint violation. Make `workflow_messages.workflow_run_id`
  nullable and drop the fallback — inbound messages created before
  a run exists stay unbound until a run consumes them. (Pre-release,
  no production callers.)
- UpsertSearchAttributesCall: every scalar was stringified (ints
  became "5", bools became "1"/"0"), which made
  SearchAttributeUpsertService's inference pick TYPE_KEYWORD for
  every input. Preserve scalar types; only strings go through
  trim/length validation.
- WorkflowSearchAttribute::inferType: the old `len <= MAX_KEYWORD`
  threshold reused the column width (255), so "a long description"
  still inferred keyword. Replace with a semantic heuristic: strings
  longer than the keyword column OR containing whitespace are
  TYPE_STRING; short identifier-like strings are TYPE_KEYWORD.

Test-infrastructure fixes:
- tests/TestCase::defineDatabaseMigrations ran `loadLaravelMigrations()`
  and then `migrate:fresh`. `migrate:fresh` drops everything and
  re-runs only registered paths — which did not include Laravel's
  defaults — so the users table got created and then dropped.
  `SerializesModelsTest` asserts against `Workbench\App\Models\User`,
  so both job-serialization-with-model tests errored. Swap the order:
  `migrate:fresh` first, `loadLaravelMigrations` after.
- tests/Unit/V2/ChildCallServiceTest: reparent from
  `Orchestra\Testbench\TestCase` to `Tests\TestCase` so the workflow
  package provider is registered and migrations run (same pattern as
  the three reparents in the first #440 pass; this one was missed).
- tests/Unit/Providers/WorkflowServiceProviderTest::
  testCarryingForwardPackageMigrationOverCustomizedPublishedSet:
  Laravel's Migrator calls `require_once` on each migration path. If
  an earlier test in the class (`testPublishedMigrationsCanRunAsInstallSource`)
  has already required the same absolute path under
  `database_path('migrations')`, editing the file in-place has no
  effect — the cached anonymous-class factory still returns the
  original Migration. Publish into a per-test scratch directory
  under `storage_path('app/test-migrations-<uniqid>')` so each test
  gets fresh paths, and also call `DB::purge()` after
  `Schema::dropAllTables()` because SQLite's `refreshDatabaseFile`
  truncates the file in place without closing the live PDO
  connection.

Test-expectation fixes (the code is correct; tests had drifted):
- MessageServiceTest: set `current_run_id` on the instance in the
  test helpers (matches production WorkflowStub behavior), reorder
  the continue-as-new test to perform the cursor advance before
  creating the continued run, and correct the correlation-id
  assertion from 2 → 4 (two sendMessage calls each produce outbound
  + inbound rows, so four share the correlation).
- MemoTest::testMemosAreNotFilterableUnlikeSearchAttributes: replace
  the `whereJsonContains('value', 'acme')` DB predicate with a
  PHP-side filter. `whereJsonContains` is not portable across
  MySQL / PostgreSQL / SQLite; the test documents a *non-filterability*
  contract and does not need a backend-specific predicate.
- SearchAttributeTest: no test changes needed — the service/model
  fixes above flip them green.
- ActivityMiddlewareTest::testMiddleware and
  WorkflowStubTest::testComplete: both relied on `Cache::flush()`
  releasing the ShouldBeUnique lock acquired by an earlier dispatch.
  That does not work: on the array driver, `flush()` clears the
  key-value store but not the separate `$locks` property; on redis
  with testbench's default config, the cache connection and the
  lock_connection resolve to different redis databases, so
  `flushdb()` on one does not touch the other. And because
  `WithoutOverlappingMiddleware` can `$job->release()` a nested
  dispatch before it runs, `CallQueuedHandler` never calls
  `ensureUniqueJobLockIsReleased` — the lock outlives the "job."
  Release the lock explicitly via `UniqueLock::release(new Job(...))`
  between dispatches.
- WithoutOverlappingMiddlewareTest::testAllowsOnlyOneWorkflowInstance:
  the RedisStore returns semaphore counts as numeric strings while
  the middleware itself `(int)`-casts on every read. Cast at the
  assertion to match.
- V2RepairPassCommandTest::testItIgnoresTheLoopThrottle...: the
  production repair already writes `signal_name` to the task payload;
  the test's expected array had drifted. Add it.

Result: `vendor/bin/phpunit --testsuite unit` against a SQLite
testbench.sqlite with `CACHE_DRIVER=redis` / `CACHE_STORE=redis`
finishes with 964/964 passing (2 pre-existing skipped, 0 errors,
0 failures). Pass a fresh (or non-existent) testbench.sqlite — a
0-byte stale file trips SQLite's "database disk image is malformed".

+++ tests/Unit/Middleware/WithoutOverlappingMiddlewareTest.php
@@ -60,7 +60,10 @@ final class WithoutOverlappingMiddlewareTest extends TestCase
             $activity,
             $middleware3
         ) {
-            $this->assertSame(1, Cache::get($middleware1->getWorkflowSemaphoreKey()));
+            // Cast because the RedisStore returns the semaphore count as a
+            // numeric string whereas the middleware itself stores it as int;
+            // the middleware's own reads cast the same way.
+            $this->assertSame(1, (int) Cache::get($middleware1->getWorkflowSemaphoreKey()));
             $this->assertNull(Cache::get($middleware1->getActivitySemaphoreKey()));
 
             $middleware2->handle($workflow2, static function ($job) {
